requirements.txt
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
eventlet==0.33.3
Flask==3.1.2
Flask-MySQL==1.6.0
Flask-Session==0.8.0
Flask-SocketIO==5.3.5
greenlet==3.2.4
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
mysql-connector-python==9.5.0
PyMySQL==1.1.2
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.10.0
simple-websocket==1.1.0
six==1.17.0
Werkzeug==3.1.4
wsproto==1.3.2
